distribute/0.0.5/build_shell/teamvision/requirements.txt
django==2.0
mysqlclient
pymysql
pillow
qrcode
httplib2
django-websocket-redis
pycrypto
coreapi
markdown
django-filter
django-crispy-forms
django-guardian
djangorestframework==3.8.2
django-url-filter
django-cors-headers
django-oauth-toolkit
mongoengine
blinker
pymongo
django-rest-framework-mongoengine
filetype
pytz
xlwt
